test: adding test for createPaginationQuery
