+++ src/components/Executions/taskExecutionQueries.ts
@@ -1,5 +1,4 @@
-import { QueryType } from 'components/data/queries';
-import { QueryInput } from 'components/data/types';
+import { QueryInput, QueryType } from 'components/data/types';
 import {
     getTaskExecution,
     listTaskExecutions,
