+++ src/components/Executions/useWorkflowExecution.ts
@@ -1,6 +1,5 @@
 import { APIContextValue, useAPIContext } from 'components/data/apiContext';
-import { QueryType } from 'components/data/queries';
-import { QueryInput } from 'components/data/types';
+import { QueryInput, QueryType } from 'components/data/types';
 import { useConditionalQuery } from 'components/hooks/useConditionalQuery';
 import { maxBlobDownloadSizeBytes } from 'components/Literals/constants';
 import {
