+++ src/components/Task/taskQueries.ts
@@ -1,5 +1,4 @@
-import { QueryType } from 'components/data/queries';
-import { QueryInput } from 'components/data/types';
+import { QueryInput, QueryType } from 'components/data/types';
 import { getTask, Identifier, TaskTemplate } from 'models';
 import { QueryClient } from 'react-query';
