+++ src/components/Workflow/workflowQueries.ts
@@ -1,5 +1,4 @@
-import { QueryType } from 'components/data/queries';
-import { QueryInput } from 'components/data/types';
+import { QueryInput, QueryType } from 'components/data/types';
 import { extractTaskTemplates } from 'components/hooks/utils';
 import { getWorkflow, Workflow, WorkflowId } from 'models';
 import { QueryClient } from 'react-query';
