--- src/components/data/queries.ts
@@ -1,13 +1,0 @@
-export enum QueryType {
-    NodeExecutionDetails = 'NodeExecutionDetails',
-    NodeExecution = 'nodeExecution',
-    NodeExecutionList = 'nodeExecutionList',
-    NodeExecutionChildList = 'nodeExecutionChildList',
-    TaskExecution = 'taskExecution',
-    TaskExecutionList = 'taskExecutionList',
-    TaskExecutionChildList = 'taskExecutionChildList',
-    TaskTemplate = 'taskTemplate',
-    Workflow = 'workflow',
-    WorkflowExecution = 'workflowExecution',
-    WorkflowExecutionList = 'workflowExecutionList'
-}
